Edit JS function and file name
