--- src/removeTask.js
@@ -1,9 +1,0 @@
-const markAsCompleted = (task) => {
-  task.completed = true;
-};
-
-const markAsIncomplete = (task) => {
-  task.completed = false;
-};
-
-export { markAsCompleted, markAsIncomplete };
